Update e2e template
